requirements.txt
streamlit
easyocr
psycopg2-binary
pandas
pillow
numpy
